Move command parsers to command modules
